Guardado de Cab y Lectura de Detalles de factura.

+++ Models/TotalImpuesto.cs
@@ -9,7 +9,7 @@ namespace ConorteClientAPI.Models
     class TotalImpuesto
     {
         public int codigo { get; set; }
-        public decimal codigoPorcentaje { get; set; }
+        public int codigoPorcentaje { get; set; }
         public decimal baseImponible { get; set; }
         public decimal  valor { get; set; }
     }
